requirements.txt
pillow
typeguard
tensorflow-gpu
requests
